+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All the following depend on Core Branding v4+.
 
 - `Button` [BREAKING] Make `aria-label` required.
 - `Icon` [BREAKING] `<i>` is for italic, not for icons. Only `<span class="..."></span>` is allowed. [WCAG 2.1 AA - Identify purpose](https://www.w3.org/TR/WCAG21/#identify-purpose)
+- `Calendar` Add `aria-label` for navigation.
 - `Badge` Add `aria-label` and `title`.
 - `Alert` Add `aria-label`.
 - `Avatar` Add `u-screen-reader-only` 'User Icon' info.